chore(transport): Clean up server and channel (#174)
* chore(transport): Clean up server and channel

* Fix tls feature compilation
